관리자 및 회원 controller 경로 변경 및 불필요한 log 삭제
