+++ src/main/java/com/edu/wing/inquiry/controller/MemberApiInquiryController.java
@@ -40,4 +40,20 @@ public class MemberApiInquiryController {
     resultMap.put("alertMsg", "문의 등록에 성공했습니다.");
     return ResponseEntity.ok().body("문의가 성공적으로 등록되었습니다");
   }
+
+  @GetMapping("/list/{inquiryNo}")
+  public ResponseEntity<Map<String, Object>> inquiryDetail(@PathVariable int inquiryNo, @RequestParam int curPage) {
+
+    Map<String, Object> resultMap = inquiryService.memberInquirySelectOne(inquiryNo);
+
+    if (resultMap == null) {
+      resultMap.put("status", "failed");
+      resultMap.put("alertMsg", "서버 오류로 인해 정보를 불러올 수 없습니다. 잠시 후 다시 시도해주세요.");
+      return ResponseEntity.badRequest().body(resultMap);
+    }
+
+    resultMap.put("curPage", curPage);
+
+    return ResponseEntity.ok().body(resultMap);
+  }
 }
