+++ src/main/java/com/edu/wing/post/controller/AdminApiPostController.java
@@ -18,16 +18,11 @@ import java.util.Map;
 @RequestMapping("admin/api/cs/post")
 public class AdminApiPostController {
 
-  private static final Logger log = LoggerFactory.getLogger(AdminApiPostController.class);
-  private static final String LOG_TITLE = "==AdminApiPostController==";
-
   @Autowired
   private PostService postService;
 
   @PostMapping("/list/add")
   public ResponseEntity<?> addPost(@RequestBody PostVo postVo, HttpSession httpSession) {
-    log.info(LOG_TITLE);
-    log.info("addPost postVo: {}", postVo);
 
     Map<String, Object> resultMap = new HashMap<>();
 
@@ -55,11 +50,57 @@ public class AdminApiPostController {
 
   }
 
+  @GetMapping("/list/{postNo}")
+  public ResponseEntity<Map<String, Object>> postDetail(@PathVariable int postNo, @RequestParam(defaultValue = "1") int curPage, @RequestParam(defaultValue = "") String postSearch,
+                                                        @RequestParam(defaultValue = "2") int noticeBoardNo) {
+
+    Map<String, Object> resultMap = new HashMap<>();
+
+    Map<String, Object> postVo = postService.postSelectOne(postNo);
+
+    if (postVo == null) {
+      resultMap.put("status", "failed");
+      resultMap.put("alertMsg", "서버 오류로 인해 정보를 불러올 수 없습니다. 잠시 후 다시 시도해 주세요.");
+
+      return ResponseEntity.badRequest().body(resultMap);
+    }
+
+    resultMap.put("curPage", curPage);
+    resultMap.put("postSearch", postSearch);
+    resultMap.put("noticeBoardNo", noticeBoardNo);
+    resultMap.put("postVo", postVo);
+
+    return ResponseEntity.ok().body(resultMap);
+  }
+
+  @GetMapping("/list/{postNo}/update")
+  public ResponseEntity<?> updatePost(@PathVariable int postNo, @RequestParam String curPage, @RequestParam(defaultValue = "") String postSearch,
+                                      @RequestParam(defaultValue = "2") int noticeBoardNo) {
+
+    Map<String, Object> resultMap = new HashMap<>();
+
+    resultMap.put("curPage", curPage);
+    resultMap.put("postSearch", postSearch);
+    resultMap.put("noticeBoardNo", noticeBoardNo);
+
+    Map<String, Object> postVo = postService.postSelectOne(postNo);
+
+    if (postVo == null) {
+      resultMap.put("status", "failed");
+      resultMap.put("alertMsg", "서버 오류가 발생하여 정보를 불러올 수 없습니다. 잠시 후 다시 시도해 주세요."); // 수정 실패에 대한 안내 메시지
+      return ResponseEntity.badRequest().body(resultMap);
+    }
+
+    resultMap.put("status", "success");
+    resultMap.put("postVo", postVo); // 수정 성공에 대한 안내 메시지
+
+    return ResponseEntity.ok().body(resultMap);
+
+  }
+
   @PatchMapping("/list/{postNo}/update")
   public ResponseEntity<?> postUpdate(@PathVariable int postNo,  @RequestBody Map<String, String> updateData, @RequestParam(defaultValue = "1") String curPage, @RequestParam(defaultValue = "") String postSearch,
                                       @RequestParam(defaultValue = "2") int noticeBoardNo, HttpSession httpSession) {
-    log.info(LOG_TITLE);
-    log.info("updatePost postNo: {}, updateData: {}, curPage: {}, postSearch: {}, noticeBoardNo: {}", postNo, updateData, curPage, postSearch, noticeBoardNo);
 
     Map<String, Object> resultMap = new HashMap<>();
 
@@ -84,8 +125,6 @@ public class AdminApiPostController {
   @DeleteMapping("/list/{postNo}/delete")
   public ResponseEntity<?> deletePost(@PathVariable int postNo, @RequestParam String curPage, @RequestParam(defaultValue = "") String postSearch,
                                       @RequestParam(defaultValue = "2") int noticeBoardNo) {
-    log.info(LOG_TITLE);
-    log.info("deletePost postNo: {}, curPage: {}, postSearch: {}", postNo, curPage, noticeBoardNo);
 
     Map<String, Object> resultMap = new HashMap<>();
