+++ src/main/resources/static/js/admin/inquiry/inquiry.js
@@ -41,8 +41,6 @@ function createDetailView(data, curPage, inquirySearch) {
   const formattedInquiryDate = formatDate(data.INQUIRYDATE);
   const formattedAnswerDate = data.ANSWERDATE ? formatDate(data.ANSWERDATE) : '';
 
-  console.log(inquirySearch);
-
   const inquiryDetail = `
     <div class="title-container one-line">
       <div id="title" class="title btn__yellow text__white">
@@ -187,7 +185,6 @@ function createUpdateView(res, curPage, inquirySearch) {
   const formattedInquiryDate = formatDate(res.INQUIRYDATE);
   const formattedAnswerDate = res.ANSWERDATE ? formatDate(res.ANSWERDATE) : '';
 
-  console.log(res)
   const inquiryCommentUpdate = `
   <div class="title-container">
     <div id="title" class="title btn__yellow text__white">
@@ -313,7 +310,6 @@ function createAddView(res, curPage, inquirySearch) {
   const formattedInquiryDate = formatDate(res.INQUIRYDATE);
   const formattedAnswerDate = res.ANSWERDATE ? formatDate(res.ANSWERDATE) : '';
 
-  console.log(res)
   const inquiryCommentAdd = `
   <div class="title-container">
     <div id="title" class="title btn__yellow text__white">
